requirements-test.txt
celery==4.2.1
django-bootstrap3==11.0.0
django-extensions==2.1.5
django-model-utils==3.1.2
flake8==3.7.5
freezegun==0.3.11
mock==2.0.0
psycopg2==2.7.7
python-dateutil==2.8.0
